Change specification for cart_less_than_stock
Change the specification for cart_less_than_stock to state that if there
exists an item in a shopping cart whose quantity exceeds that of the
stock, then checking out the cart should fail.

We need to tweak the theorem this way because the original specification
did not relate the contents of the cart to the checkout function at all,
making it impossible to prove theorems about the cart's contents.